Allow PCEAS to build on Windows with VS2019 by adding xgetopt
See https://github.com/matthklo/xgetopt

Also fix a bunch of VS2019 compiler warnings in the PCEAS code.

+++ src/mkit/as/assemble.c
@@ -435,7 +435,7 @@ addinst(struct t_opcode *optbl)
 	while (optbl->name) {
 		/* calculate instruction hash value */
 		hash = 0;
-		len = strlen(optbl->name);
+		len = (int)strlen(optbl->name);
 		ptr = optbl->name;
 
 		for (i = 0; i < len; i++) {
